requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-docx==0.8.11
Pillow==10.0.1
youtube-dl==2021.12.17
yt-dlp==2023.7.6
selenium==4.15.0
webdriver-manager==4.0.1
flask-cors==4.0.0
